cmd/tier: speed up telemetery reporting (#61)
This commit changes event reporting to the telemetry server in a way to
prevents the end-of-process lag previously seen. It accomplishes this by
warming up an http/2 connection and consequentially, warming up the
server if it was asleep. It does this in the background to race the
final step in the running of tier so that when events need to be
written, they will not have to wait on any setup or warmup before
getting their events out.

Before (after binary warmup):

    ; time tier version
    0.1.0-dev
      go version: go1.19

    real    0m0.306s
    user    0m0.009s
    sys     0m0.010s

After:

    ; time tier version
    0.1.0-dev
      go version: go1.19

    real    0m0.043s
    user    0m0.007s
    sys     0m0.008s