cmd/tier: add title field to feature, and other PJ changes (#66)
This commit changes PricingJSON in the following ways:

* Adds a title field to the feature object
* Moves the interval field up from feature to plan
* Moves the currency field up from feature to plan

The above changes were made to support a new plan name (for display)
format in Stripe. The new format is now:

  "PlanName (Interval): FeatureName (PlanVersion)"

In addition to the above change, this commit also contains a
rip-and-replace of the Stripe SDK with a much simpler recipe of the
standard http client, the json package, and a few helpers for building
Stripe form-encoded request payloads and retrieving lists of objects
back from stripe. This resulting in a significant reduction in the
codebase that dealt with Stripe and gives us a lot more flexibility and
power now, and in the future.