cmd/tier: stop sending RST_STREAM to upstream telemetry server (#68)
Previously canceling the request before exit, despite the request
payload being written, caused the http2 client the send RST_STREAM to
the upstream server (as expected), however, it seems the upstream server
(in this case: Vercel/lambda) will give up on sending the request to the
function if it encounters a RST_STREAM for some reason. The ordering that
causes the behavior isn't clear to me at the moment, but preventing the
RST_STREAM frame from being sent allows us to fire-and-forget with a
high chance the payload makes it to the function in the face of the
process exiting and cutting off the connection to the upstream server.