cmd/tier: fix subscribe and add phases
This fixes the subscribe command so that it updates the current phase
instead of appending a phase at the end of the existing phases, which
lead to hitting the 10 phase limit.

This commit also adds the phases command for listing an orgs phases.

This commit also introduces added testing and fixes some minor issues
regarding Stripe idempotency keys not being wiped along with test data.