api: change /v1/subscribe default behavior
This changes the /v1/subscribe default behavior to match the CLI's
subscribe command. It takes only a single plan, no effective date, and
updates the current phase in a scheudule with the new features in the
provided plan. There remains a path to allow clients to do more advanced
things like specify multi-phase schedules and update-in-place or
attached to the current phase in interesting ways use the Current field
on tier.Phase.

Fixes #93