cmd/tier: update --live flag rules
The --live flag is now required if key found in the environment is also
a live key; otherwise it is not allowed to be set.

This means the --live flag is now not only a way to specify one wants to
work in live mode, but also a confirmation that live mode is intended if
STRIPE_API_KEY is set with a test mode key.