api: add /v1/phase (#113)
This adds the new endpoint /v1/phase which takes an org as a parameter
and reports the current phase's features, plans, and their versions. As
an added bonus, it also returns the computed "fragments" for the phase.
Fragments are features in the phase that are not accompanied by their
siblings, or features without plans.