all: make features have versions
This changes the internal understanding and representiation of a "plan"
by moving it from a seperate field to being included in the feature IDs,
which now have versions. Tier now groups features by their versions iif
their versions start with the prefix ("plan:").

Work will continue after this to further reduce the complexity that
keeping plans and features seperate was causing.