cmd/tier: require the explicit intention to use stdin (#144)
Previously tier push was trying to be smart and read from stdin based on
a handful of heuristics, such as if stdin was a TTY, or if the filename
was the ("-"), etc. This replaces the need for all of that by requiring
the user explicitly initiates the reading of stdin using the special
filename ("-").

This also fixes bugs with exit codes and help messages.