cmd/tier: display account specific URLs on push
This fixes a confusing bug that sent users to their default, logged-in
Stripe account even if they were using an isolated account with the
switch command.

The links displayed are now relative to the account the features were
pushed to.

This commit also reintroduces deep linking to prices pushed.

Fixes #166