api,cmd/tier: accept extra org info on schedule (#179)
This commit makes subscribe accept additional information about an org to
attach to it when creating or updating schedules in stripe. The new
information is:

- description
- name
- phone
- email
- metadata

Metadata containing any key prefixed with ("tier.") is an error and the
the whole request is rejected.

This commit also makes /v1/whois respond with all up-to-date information
attached to an orgs underlying Stripe customer if the include=info query
param is present. The response including info does not expose any
metadata added by Tier.