api: support free trials (#203)
This commit adds a Trial field to the Phase object accepted by the API
for creating and updating subscriptions and schedules. The new Trial
field tell Stripe to count the Phase as a trial phase, which Stripe will
honor by counting all features as free during the lifetime of the Phase.

This also adds the --trial=<days> flag to the CLI subscribe subcommand
for specifying trial periods. If days is negative, then a single phase
is scheduled as a trial that never ends. If days is zero, a single
non-trial phase is scheduled with the features/plans provided. If days
is positive, a trial phase is scheduled ending after the number of days
provided, followed by a phase that never ends for the features/plans
provided.