control: rework subscription and schedule approach (#218)
This changes the scheduling logic to consider the subscription as the source of
truth, rather than trying to force the schedule API into being it. The new
approach drastically simplifies the ability to reason about how Tier converts a
single timeline into the complex beast that is Stripe schedules and
subscription schedules because we can now say that: All schedule reads are done
through subscriptions via the active subscription's current "schedule," and all
schedules are created and updated via schedules.

Create: Schedule API creates the subscription.

Updates: The schedule API updates the subscription.

Reads: The subscription API shows the current state, including the current
current subscription schedule

This also relaxes constraints for canceling phases. Previous to this
commitment, cancellations have to go into effect immediately. Now cancel phases
may appear in multi-phase schedules where the canceling happens in the future.
The new restriction is that they are the last phase.

The new "last phase" restriction may be relaxed in the future by canceling the
current schedule as it does now but then creating a new subscription schedule,
scheduled in the future.

This also adds the ability to update default payment information for a stripe.
customer via OrgInfo.