control: return pseudo phases if no schedule (#226)
This prevents lookupPhases from passing empty schedule IDs to Stripe,
which causes an API error that bubbles up to Tier clients.

The lookupPhases method will now return pseudo Phases in place of
schedule phases if no schedule exists for a subscription at the time of
calling. If a trialing, the first phase will be the trial phase. If a
cancel is scheduled, a final "cancel" phase will be appended after the
paid phase, if any.

This allows callers to think in Phases/Schedules only and not
worry about what a subscription is as well.

Fixes #221