use "subscription.status" instead of trial_end. (#231)
The trial_end field returned by stripe is always non-zero if there was a
trial, ever. This means trial_end is not reset after any transition out
of trialing.

Also, account for canceled_at status. This means lookup phases will now
automatically add a cancel phase to the pseudo phases, effective at
canceled_at. If there is a cancel_at and the subscription is yet to be
canceled, then use that as the start of the cancel phase, but mark it as
not current.