control: convert single tier to metered+per_unit (#267)
This changes how Tier interprets and converts tiers to Stripe prices.

Previously, a single tier in a pricing.json would convert to a metered
price with one or two tiers depending on if the tier had an inf limit or
not. If it did not have an inf limit, a catch-all was created to satisfy
Stripe. This produced awkward invoices.

This commit changes Tier to convert a feature that has a single tier
without a base price to be a metered+per_unit price in Stripe, which
shows up less awkwardly on invoices.

This also opens us up to support pricing transforms in Stripe for these
types of prices.