control: be more aware of test clocks
This commit fixes a bug with test clocks where customer IDs were cached
based only on the org ID, not by test clocks. This resulted in invalid
cache hits due to org name collisions if the same org name was used
across test clocks because the first customer ID would be cached and
then used in requests targeting different clocks where the customer ID
did not exist.

This also narrows the list of returned customers by Stripe to those in a
test clock, if any.