control: fix bug when scheduling from subscription (#297)
Previously, when scheduling from a subscription, we would send 'now'
as the start of the first phase, which Stripe would interpret as "update
current period start date to 'now'", which is illegal. Instead, we now
send the current period start date as the start of the first phase we
get back after creating the schedule from the subscription.